[LIB-637] add rename mixin for the rename functionality, add this to the ResponseTemplate models;
