+++ syncano/models/instances.py
@@ -2,9 +2,10 @@
 
 from . import fields
 from .base import Model
+from .mixins import RenameMixin
 
 
-class Instance(Model):
+class Instance(RenameMixin, Model):
     """
     OO wrapper around instances `link <http://docs.syncano.com/docs/getting-started-with-syncano#adding-an-instance>`_.
 
@@ -74,20 +75,6 @@ class Instance(Model):
                 'path': '/v1.1/instances/',
             }
         }
-
-    def rename(self, new_name):
-        """
-        A method for changing the instance name;
-
-        :param new_name: the new name for the instance;
-        :return: a populated Instance object;
-        """
-        rename_path = self.links.rename
-        data = {'new_name': new_name}
-        connection = self._get_connection()
-        response = connection.request('POST', rename_path, data=data)
-        self.to_python(response)
-        return self
 
 
 class ApiKey(Model):
