+++ tests/integration_test_reponse_templates.py
@@ -49,3 +49,13 @@ class ResponseTemplateApiTest(InstanceMixin, IntegrationTest):
 
         rendered = render_template.render(context={'objects': [3]})
         self.assertEqual(rendered, '<li>3</li>')
+
+    def test_rename(self):
+        name = 'some_old_new_name_for_template'
+        new_name = 'some_new_name_for_template'
+
+        template = ResponseTemplate.please.create(name=name, content='<div></div>', content_type='text/html',
+                                                  context={'two': 2})
+        template = template.rename(new_name=new_name)
+
+        self.assertEqual(template.name, new_name)
